requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
altair==5.2.0
antlr4-python3-runtime==4.9.3
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.7.0
bokeh==2.4.3
boto3==1.33.9
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
et-xmlfile==1.1.0
fastapi==0.105.0
filelock==3.12.4
fire==0.5.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.12.1
gast==0.5.4
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.62.0
grpcio==1.59.0
gTTS==2.4.0
h5py==3.9.0
humanfriendly==10.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.11.0
ipykernel==6.25.2
ipython==8.16.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.14.0
kiwisolver==1.4.5
langchain==0.0.350
langchain-core==0.1.0
langchain-experimental==0.0.47
lazy_loader==0.3
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.14.1
openai==1.3.8
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.0.1
pip==23.2.1
platformdirs==3.10.0
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==3.20.2
psutil==5.9.5
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
pypdf==3.17.3
pypdfium2==4.25.0
PyPika==0.48.9
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
pytorch-pretrained-bert==0.6.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.5.2
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
s3transfer==0.8.2
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
setuptools==65.5.0
simpleaudio==1.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
streamlit==1.29.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.12
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchfile==0.1.0
torchvision==0.16.2+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.35.2
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.5
unstructured-client==0.15.0
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.8
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.0
wheel==0.41.2
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.9
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0

